Refactoring, add fit_airlines_card.dart
